legal-rag-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
faiss-cpu
sentence-transformers
pydantic
langdetect